+++ src/main/java/mcjty/lib/gui/widgets/BlockRender.java
@@ -145,7 +145,7 @@ public class BlockRender extends AbstractWidget<BlockRender> {
     }
 
     @Override
-    public Widget<?> mouseClick(int x, int y, int button) {
+    public Widget<?> mouseClick(double x, double y, int button) {
         if (isEnabledAndVisible()) {
             fireSelectionEvents();
             long t = System.currentTimeMillis();
